feat: add upload author feature on firebase and get data correctly
