+++ plugins/firebase.js
@@ -1,5 +1,6 @@
 import { initializeApp } from 'firebase/app'
 import { getStorage } from 'firebase/storage'
+import { getFirestore } from 'firebase/firestore'
 
 const firebaseConfig = {
   apiKey: 'AIzaSyBEFVYqJF4gar7WrhscA9YDizWyRi9LBgI',
@@ -13,3 +14,4 @@ const firebaseConfig = {
 
 const app = initializeApp(firebaseConfig)
 export const storage = getStorage(app)
+export const db = getFirestore(app)
