+++ store/getters.js
@@ -1,0 +1,8 @@
+export default {
+  loadedPosts(state) {
+    return state.loadedPosts
+  },
+  isAuthenticated(state) {
+    return state.token != null
+  },
+}
