+++ store/state.js
@@ -1,0 +1,4 @@
+export default () => ({
+  loadedPosts: [],
+  token: null,
+})
